requirements.txt
Django==2.2
django-extensions==2.1.6
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
whitenoise==3.2
